+++ packages/dodo-ui-date/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch Changes
 
+- DatePicker trigger button hide added
 - Updated dependencies
   - @flightlesslabs/dodo-ui@0.32.0